Consolidate right component display name with mui component name
