+++ lib/src/_shared/withDefaultProps.tsx
@@ -3,10 +3,11 @@ import getThemeProps from '@material-ui/styles/getThemeProps';
 import { useTheme } from '@material-ui/core/styles';
 
 export function withDefaultProps<T>(
-  { name }: { name: string },
+  { name }: { name: string; skipDisplayName?: boolean },
   Component: React.ComponentType<T>
 ): React.FC<T> {
-  return ({ ...props }) => {
+  const componentName = name.replace('Mui', '');
+  const WithDefaultProps = (props: T) => {
     const theme = useTheme();
     const propsWithDefault = getThemeProps<any, T, string>({
       props,
@@ -14,6 +15,11 @@ export function withDefaultProps<T>(
       name,
     });
 
+    Component.displayName = componentName;
+
     return <Component {...propsWithDefault} />;
   };
+
+  WithDefaultProps.displayName = `WithDefaultProps(${componentName})`;
+  return WithDefaultProps;
 }
